compute thermo is in md namespace
