+++ hoomd/__init__.py
@@ -40,7 +40,6 @@ from hoomd import cite
 from hoomd import analyze
 from hoomd import benchmark
 from hoomd import comm
-from hoomd import compute
 from hoomd import dump
 from hoomd import init
 from hoomd import integrate
