+++ hoomd/md/ComputeThermoGPU.h
