+++ hoomd/md/ComputeThermoHMAGPU.cc
@@ -2,21 +2,22 @@
 // This file is part of the HOOMD-blue project, released under the BSD 3-Clause License.
 
 
-// Maintainer: joaander
+// Maintainer: ajs42
 
-/*! \file ComputeThermoGPU.cc
-    \brief Contains code for the ComputeThermoGPU class
+/*! \file ComputeThermoHMAGPU.cc
+    \brief Contains code for the ComputeThermoHMAGPU class
 */
 
-#include "ComputeThermoGPU.h"
-#include "ComputeThermoGPU.cuh"
-#include "GPUPartition.cuh"
+
+#include "ComputeThermoHMAGPU.h"
+#include "ComputeThermoHMAGPU.cuh"
+#include "hoomd/GPUPartition.cuh"
 
 namespace py = pybind11;
 
 #ifdef ENABLE_MPI
-#include "Communicator.h"
-#include "HOOMDMPI.h"
+#include "hoomd/Communicator.h"
+#include "hoomd/HOOMDMPI.h"
 #endif
 
 #include <iostream>
@@ -24,35 +25,52 @@ using namespace std;
 
 /*! \param sysdef System for which to compute thermodynamic properties
     \param group Subset of the system over which properties are calculated
+    \param temperature The temperature that governs sampling of the integrator
+    \param harmonicPressure The contribution to the pressure from harmonic fluctuations
     \param suffix Suffix to append to all logged quantity names
 */
 
-ComputeThermoGPU::ComputeThermoGPU(std::shared_ptr<SystemDefinition> sysdef,
-                                   std::shared_ptr<ParticleGroup> group,
-                                   const std::string& suffix)
-    : ComputeThermo(sysdef, group, suffix), m_scratch(m_exec_conf), m_scratch_pressure_tensor(m_exec_conf),
-        m_scratch_rot(m_exec_conf)
+ComputeThermoHMAGPU::ComputeThermoHMAGPU(std::shared_ptr<SystemDefinition> sysdef,
+                                   std::shared_ptr<ParticleGroup> group, const double temperature,
+                                   const double harmonicPressure, const std::string& suffix)
+    : ComputeThermoHMA(sysdef, group, temperature, harmonicPressure, suffix), m_scratch(m_exec_conf)
     {
     if (!m_exec_conf->isCUDAEnabled())
         {
-        m_exec_conf->msg->error() << "Creating a ComputeThermoGPU with no GPU in the execution configuration" << endl;
-        throw std::runtime_error("Error initializing ComputeThermoGPU");
+        m_exec_conf->msg->error() << "Creating a ComputeThermoHMAGPU with no GPU in the execution configuration"
+                                  << endl;
+        throw std::runtime_error("Error initializing ComputeThermoHMAGPU");
         }
 
-    m_block_size = 256;
+    m_block_size = 512;
 
-    hipEventCreateWithFlags(&m_event, hipEventDisableTiming);
+    if (m_exec_conf->allConcurrentManagedAccess())
+        {
+        auto gpu_map = m_exec_conf->getGPUIds();
+
+        // set up GPU memory mappings
+        for (unsigned int idev = 0; idev < m_exec_conf->getNumActiveGPUs(); ++idev)
+            {
+            // only optimize access for those fields used in force computation
+            // (i.e. no net_force/virial/torque, also angmom and inertia are only used by the integrator)
+            cudaMemAdvise(m_lattice_site.get(), sizeof(Scalar3)*m_lattice_site.getNumElements(),
+                          cudaMemAdviseSetAccessedBy, gpu_map[idev]);
+            }
+        CHECK_CUDA_ERROR();
+        }
+
+    cudaEventCreate(&m_event, cudaEventDisableTiming);
     }
 
 //! Destructor
-ComputeThermoGPU::~ComputeThermoGPU()
+ComputeThermoHMAGPU::~ComputeThermoHMAGPU()
     {
-    hipEventDestroy(m_event);
+    cudaEventDestroy(m_event);
     }
 
 /*! Computes all thermodynamic properties of the system in one fell swoop, on the GPU.
  */
-void ComputeThermoGPU::computeProperties()
+void ComputeThermoHMAGPU::computeProperties()
     {
     // just drop out if the group is an empty group
     if (m_group->getNumMembersGlobal() == 0)
@@ -71,12 +89,9 @@ void ComputeThermoGPU::computeProperties()
     unsigned int old_size = m_scratch.size();
 
     m_scratch.resize(num_blocks);
-    m_scratch_pressure_tensor.resize(num_blocks*6);
-    m_scratch_rot.resize(num_blocks);
 
     if (m_scratch.size() != old_size)
         {
-        #ifdef __HIP_PLATFORM_NVCC__
         if (m_exec_conf->allConcurrentManagedAccess())
             {
             auto& gpu_map  = m_exec_conf->getGPUIds();
@@ -84,31 +99,25 @@ void ComputeThermoGPU::computeProperties()
             // map scratch array into memory of all GPUs
             for (unsigned int idev = 0; idev < m_exec_conf->getNumActiveGPUs(); ++idev)
                 {
-                cudaMemAdvise(m_scratch.get(), sizeof(Scalar4)*m_scratch.getNumElements(), cudaMemAdviseSetAccessedBy, gpu_map[idev]);
-                cudaMemAdvise(m_scratch_pressure_tensor.get(), sizeof(Scalar)*m_scratch_pressure_tensor.getNumElements(), cudaMemAdviseSetAccessedBy, gpu_map[idev]);
-                cudaMemAdvise(m_scratch_rot.get(), sizeof(Scalar)*m_scratch_rot.getNumElements(), cudaMemAdviseSetAccessedBy, gpu_map[idev]);
+                cudaMemAdvise(m_scratch.get(), sizeof(Scalar3)*m_scratch.getNumElements(), cudaMemAdviseSetAccessedBy,
+                              gpu_map[idev]);
                 }
             CHECK_CUDA_ERROR();
             }
-        #endif
 
         // reset to zero, to be on the safe side
-        ArrayHandle<Scalar4> d_scratch(m_scratch, access_location::device, access_mode::overwrite);
-        ArrayHandle<Scalar> d_scratch_pressure_tensor(m_scratch_pressure_tensor, access_location::device, access_mode::overwrite);
-        ArrayHandle<Scalar> d_scratch_rot(m_scratch_rot, access_location::device, access_mode::overwrite);
+        ArrayHandle<Scalar3> d_scratch(m_scratch, access_location::device, access_mode::overwrite);
 
-        hipMemset(d_scratch.data, 0, sizeof(Scalar4)*m_scratch.size());
-        hipMemset(d_scratch_pressure_tensor.data, 0, sizeof(Scalar)*m_scratch_pressure_tensor.size());
-        hipMemset(d_scratch_rot.data, 0, sizeof(Scalar)*m_scratch_rot.size());
+        cudaMemset(d_scratch.data, 0, sizeof(Scalar3)*m_scratch.size());
         }
 
     // access the particle data
-    ArrayHandle<Scalar4> d_vel(m_pdata->getVelocities(), access_location::device, access_mode::read);
+    ArrayHandle<Scalar4> d_pos(m_pdata->getPositions(), access_location::device, access_mode::read);
     ArrayHandle<unsigned int> d_body(m_pdata->getBodies(), access_location::device, access_mode::read);
     ArrayHandle<unsigned int> d_tag(m_pdata->getTags(), access_location::device, access_mode::read);
+    ArrayHandle<Scalar3> d_lattice_site(m_lattice_site, access_location::device, access_mode::read);
+    ArrayHandle<int3> d_image(m_pdata->getImages(), access_location::device, access_mode::read);
     BoxDim box = m_pdata->getGlobalBox();
-
-    PDataFlags flags = m_pdata->getFlags();
 
     { // scope these array handles so they are released before the additional terms are added
     // access the net force, pe, and virial
@@ -116,12 +125,7 @@ void ComputeThermoGPU::computeProperties()
     const GlobalArray< Scalar >& net_virial = m_pdata->getNetVirial();
     ArrayHandle<Scalar4> d_net_force(net_force, access_location::device, access_mode::read);
     ArrayHandle<Scalar> d_net_virial(net_virial, access_location::device, access_mode::read);
-    ArrayHandle<Scalar4> d_orientation(m_pdata->getOrientationArray(), access_location::device, access_mode::read);
-    ArrayHandle<Scalar4> d_angmom(m_pdata->getAngularMomentumArray(), access_location::device, access_mode::read);
-    ArrayHandle<Scalar3> d_inertia(m_pdata->getMomentsOfInertiaArray(), access_location::device, access_mode::read);
-    ArrayHandle<Scalar4> d_scratch(m_scratch, access_location::device, access_mode::overwrite);
-    ArrayHandle<Scalar> d_scratch_pressure_tensor(m_scratch_pressure_tensor, access_location::device, access_mode::overwrite);
-    ArrayHandle<Scalar> d_scratch_rot(m_scratch_rot, access_location::device, access_mode::overwrite);
+    ArrayHandle<Scalar3> d_scratch(m_scratch, access_location::device, access_mode::overwrite);
     ArrayHandle<Scalar> d_properties(m_properties, access_location::device, access_mode::overwrite);
 
     // access the group
@@ -130,39 +134,31 @@ void ComputeThermoGPU::computeProperties()
     m_exec_conf->beginMultiGPU();
 
     // build up args list
-    compute_thermo_args args;
+    compute_thermo_hma_args args;
     args.n_blocks = num_blocks;
     args.d_net_force = d_net_force.data;
     args.d_net_virial = d_net_virial.data;
-    args.d_orientation = d_orientation.data;
-    args.d_angmom = d_angmom.data;
-    args.d_inertia = d_inertia.data;
     args.virial_pitch = net_virial.getPitch();
-    args.ndof = m_group->getTranslationalDOF();
     args.D = m_sysdef->getNDimensions();
     args.d_scratch = d_scratch.data;
-    args.d_scratch_pressure_tensor = d_scratch_pressure_tensor.data;
-    args.d_scratch_rot = d_scratch_rot.data;
     args.block_size = m_block_size;
     args.external_virial_xx = m_pdata->getExternalVirial(0);
-    args.external_virial_xy = m_pdata->getExternalVirial(1);
-    args.external_virial_xz = m_pdata->getExternalVirial(2);
     args.external_virial_yy = m_pdata->getExternalVirial(3);
-    args.external_virial_yz = m_pdata->getExternalVirial(4);
     args.external_virial_zz = m_pdata->getExternalVirial(5);
     args.external_energy = m_pdata->getExternalEnergy();
+    args.temperature = m_temperature;
+    args.harmonicPressure = m_harmonicPressure;
 
     // perform the computation on the GPU(s)
-    gpu_compute_thermo_partial( d_properties.data,
-                        d_vel.data,
+    gpu_compute_thermo_hma_partial(d_pos.data,
+                        d_lattice_site.data,
+                        d_image.data,
                         d_body.data,
                         d_tag.data,
                         d_index_array.data,
                         group_size,
                         box,
                         args,
-                        flags[pdata_flag::pressure_tensor],
-                        flags[pdata_flag::rotational_kinetic_energy],
                         m_group->getGPUPartition());
 
     if(m_exec_conf->isCUDAErrorCheckingEnabled())
@@ -172,16 +168,13 @@ void ComputeThermoGPU::computeProperties()
     m_exec_conf->endMultiGPU();
 
     // perform the computation on GPU 0
-    gpu_compute_thermo_final( d_properties.data,
-                        d_vel.data,
+    gpu_compute_thermo_hma_final( d_properties.data,
                         d_body.data,
                         d_tag.data,
                         d_index_array.data,
                         group_size,
                         box,
-                        args,
-                        flags[pdata_flag::pressure_tensor],
-                        flags[pdata_flag::rotational_kinetic_energy]);
+                        args);
 
     if(m_exec_conf->isCUDAErrorCheckingEnabled())
         CHECK_CUDA_ERROR();
@@ -195,11 +188,13 @@ void ComputeThermoGPU::computeProperties()
     if (m_prof) m_prof->pop(m_exec_conf);
     }
 
-void export_ComputeThermoGPU(py::module& m)
+void export_ComputeThermoHMAGPU(py::module& m)
     {
-    py::class_<ComputeThermoGPU, ComputeThermo, std::shared_ptr<ComputeThermoGPU> >(m,"ComputeThermoGPU")
+    py::class_<ComputeThermoHMAGPU, ComputeThermoHMA, std::shared_ptr<ComputeThermoHMAGPU> >(m,"ComputeThermoHMAGPU")
     .def(py::init< std::shared_ptr<SystemDefinition>,
          std::shared_ptr<ParticleGroup>,
+         const double,
+         const double,
          const std::string& >())
         ;
     }
