+++ hoomd/md/ComputeThermoHMAGPU.h
