+++ hoomd/md/__init__.py
@@ -16,6 +16,7 @@ classes will not require any modifications. **Maintainer:** Joshua A. Anderson
 from hoomd.md import angle
 from hoomd.md import bond
 from hoomd.md import charge
+from hoomd.md import compute
 from hoomd.md import constrain
 from hoomd.md import dihedral
 from hoomd.md import external
