+++ hoomd/md/pytest/CMakeLists.txt
@@ -4,6 +4,7 @@ set(files __init__.py
     test_flags.py
     test_pair.py
     test_methods.py
+    test_thermo.py
     forces_and_energies.json
     )
