+++ hoomd/md/test/test_enforce2d_updater.cc
@@ -14,7 +14,7 @@
 #include "hoomd/md/AllPairPotentials.h"
 #include "hoomd/md/NeighborListBinned.h"
 #include "hoomd/md/TwoStepNVTMTK.h"
-#include "hoomd/ComputeThermo.h"
+#include "hoomd/md/ComputeThermo.h"
 #include "hoomd/filter/ParticleFilterAll.h"
 
 #ifdef ENABLE_HIP
