+++ hoomd/md/test/test_temp_rescale_updater.cc
@@ -9,12 +9,12 @@
 
 #include <memory>
 
-#include "hoomd/ComputeThermo.h"
+#include "hoomd/md/ComputeThermo.h"
 #include "hoomd/md/TempRescaleUpdater.h"
 #include "hoomd/filter/ParticleFilterAll.h"
 
 #ifdef ENABLE_HIP
-#include "hoomd/ComputeThermoGPU.h"
+#include "hoomd/md/ComputeThermoGPU.h"
 #endif
 
 #include <math.h>
