+++ hoomd/module.cc
@@ -15,8 +15,6 @@
 #include "GetarInitializer.h"
 #include "GSDReader.h"
 #include "Compute.h"
-#include "ComputeThermo.h"
-#include "ComputeThermoHMA.h"
 #include "CellList.h"
 #include "CellListStencil.h"
 #include "ForceCompute.h"
@@ -53,9 +51,7 @@
 #ifdef ENABLE_HIP
 #include <hip/hip_runtime.h>
 #include "CellListGPU.h"
-#include "ComputeThermoGPU.h"
 #include "SFCPackTunerGPU.h"
-#include "ComputeThermoHMAGPU.h"
 #endif
 
 // include MPI classes
@@ -370,8 +366,6 @@ PYBIND11_MODULE(_hoomd, m)
 
     // computes
     export_Compute(m);
-    export_ComputeThermo(m);
-    export_ComputeThermoHMA(m);
     export_CellList(m);
     export_CellListStencil(m);
     export_ForceCompute(m);
@@ -380,8 +374,6 @@ PYBIND11_MODULE(_hoomd, m)
 
 #ifdef ENABLE_HIP
     export_CellListGPU(m);
-    export_ComputeThermoGPU(m);
-    export_ComputeThermoHMAGPU(m);
 #endif
 
     // analyzers
