+++ hoomd/pytest/CMakeLists.txt
@@ -18,7 +18,6 @@ set(files __init__.py
           test_state.py
           test_simulation.py
           test_variant.py
-          test_thermo.py
           pytest-openmpi.sh
     )
